Refactor the automated tests and structure them in a better way
